Update Checkers-test.py
Added tests that check for a specific bug that I found and fixed earlier where a win would be triggered because the consecutive count would wrap around rows and columns. I find believe its good practice to write a specific unit test for bugs that you have found and fixed.